+++ test/utils/ObjectsTest.ts
@@ -37,6 +37,9 @@ var numsValues: number[] = [];
 } ());
 
 
+function dateA() { return new Date(1999, 5, 4, 3, 2, 1, 0); }
+
+
 QUnit.module("Objects", {
 });
 
@@ -112,14 +115,16 @@ QUnit.test("cloneMap", function coalesceTest(sr) {
     var res1 = Objects.cloneMap({ sub: 123, sup: 312, arc: "a" }, null, (prop) => prop.toString().substr(0, 2));
     sr.deepEqual(res1, { sub: "12", sup: "31", arc: "a" });
 
-    var res2 = Objects.cloneMap({ a: {}, b: "ab", c: true }, null, (prop) => prop.toString());
-    sr.deepEqual(res2, { a: "[object Object]", b: "ab", c: "true" });
+    var res2 = Objects.cloneMap({ a: {}, b: 2.4, c: true }, null, (prop) => typeof prop === "number" ? prop * 2 : prop.toString());
+    sr.deepEqual(res2, { a: "[object Object]", b: 4.8, c: "true" });
 
-    var res3 = Objects.cloneMap({ a: {}, b: "ab", c: true, d: 1010 }, ["a", "c"], (prop) => JSON.parse(JSON.stringify(prop)));
-    sr.deepEqual(res3, { a: {}, c: true });
+    var objA = {};
+    var res3 = <any>Objects.cloneMap({ a: objA, b: false, c: 4.8, e: true }, ["a", "c", "e"], (prop) => prop);
 
-    var res4 = Objects.cloneMap({ });
-    sr.deepEqual(res4, { });
+    sr.deepEqual(res3, { a: objA, c: 4.8, e: true });
+
+    var res4 = Objects.cloneMap(objA);
+    sr.deepEqual(res4, objA);
 });
 
 
@@ -129,7 +134,7 @@ QUnit.test("clone", function getPropTest(sr) {
         b: { 1: 1, 2: 2, 3: 3 },
         c: {
             c1: [1, 2, 4, 8],
-            c2: { c2a: [{ sup: "alpha" }, { sub: { text: "beta", values: ["b", "e", "t", "a"] } }] },
+            c2: { c2a: [{ sup: "alpha" }, { sub: { text: true, values: [false, 1.3, "t", "a"] } }] },
             c3: ["1-dimension", ["2", "dimension"], ["3", "-", "dimension"]],
         },
     };
@@ -138,7 +143,7 @@ QUnit.test("clone", function getPropTest(sr) {
         b: { 1: 1, 2: 2, 3: 3 },
         c: {
             c1: [1, 2, 4, 8],
-            c2: { c2a: [{ sup: "alpha" }, { sub: { text: "beta", values: ["b", "e", "t", "a"] } }] },
+            c2: { c2a: [{ sup: "alpha" }, { sub: { text: true, values: [false, 1.3, "t", "a"] } }] },
             c3: ["1-dimension", ["2", "dimension"], ["3", "-", "dimension"]],
         },
     };
@@ -147,7 +152,7 @@ QUnit.test("clone", function getPropTest(sr) {
         b: { 1: 1, 2: 2, 3: 3 },
         c: {
             c1: [1, 2, 4, 8],
-            c2: { c2a: [{ sup: "alpha" }, { sub: { text: "beta", values: ["b", "e", "t", "a"] } }] },
+            c2: { c2a: [{ sup: "alpha" }, { sub: { text: true, values: [false, 1.3, "t", "a"] } }] },
             c3: ["1-dimension", ["2", "dimension"], ["3", "-", "dimension"]],
         },
     };
@@ -156,48 +161,61 @@ QUnit.test("clone", function getPropTest(sr) {
     src1a.a = null;
     sr.deepEqual(res1, src1);
 
-    var res1 = Objects.clone(src1b);
+    var res2 = Objects.clone(src1b);
     src1b.c.c1 = null;
-    sr.notDeepEqual(res1, src1);
+    sr.notDeepEqual(res2, src1);
 });
 
 
 QUnit.test("cloneDeep", function getPropTest(sr) {
     var src1 = {
         a: "A",
-        b: { 1: 1, 2: 2, 3: 3 },
+        b: { 1: 1, 2: 2, 3: 3, date: dateA() },
         c: {
             c1: [1, 2, 4, 8],
-            c2: { c2a: [{ sup: "alpha" }, { sub: { text: "beta", values: ["b", "e", "t", "a"] } }] },
+            c2: { c2a: [{ sup: "alpha" }, { sub: { text: "beta", values: [false, 1.3, "t", "a"] } }] },
             c3: ["1-dimension", ["2", "dimension"], ["3", "-", "dimension"]],
         },
     };
     var src1a = {
         a: "A",
-        b: { 1: 1, 2: 2, 3: 3 },
+        b: { 1: 1, 2: 2, 3: 3, date: dateA() },
         c: {
             c1: [1, 2, 4, 8],
-            c2: { c2a: [{ sup: "alpha" }, { sub: { text: "beta", values: ["b", "e", "t", "a"] } }] },
+            c2: { c2a: [{ sup: "alpha" }, { sub: { text: "beta", values: [false, 1.3, "t", "a"] } }] },
             c3: ["1-dimension", ["2", "dimension"], ["3", "-", "dimension"]],
         },
     };
     var src1b = {
         a: "A",
-        b: { 1: 1, 2: 2, 3: 3 },
+        b: { 1: 1, 2: 2, 3: 3, date: dateA() },
         c: {
             c1: [1, 2, 4, 8],
-            c2: { c2a: [{ sup: "alpha" }, { sub: { text: "beta", values: ["b", "e", "t", "a"] } }] },
+            c2: { c2a: [{ sup: "alpha" }, { sub: { text: "beta", values: [false, 1.3, "t", "a"] } }] },
             c3: ["1-dimension", ["2", "dimension"], ["3", "-", "dimension"]],
         },
     };
 
     var res1 = Objects.cloneDeep(src1a);
+    var res2 = Objects.cloneDeep(src1b);
+
+    // check copied dates, first modify the original object dates, then ensure the copies match the original
+    src1a.b.date.setUTCFullYear(2025, 11, 25);
+    src1b.b.date.setUTCFullYear(2025, 11, 25);
+
+    sr.equal(res1.b.date.toISOString(), src1.b.date.toISOString());
+    sr.equal(res2.b.date.toISOString(), src1.b.date.toISOString());
+
+    // convert the dates to timestamps since deepEqual doesn't compare dates properly
+    res1.b.date = <any>res1.b.date.toISOString();
+    res2.b.date = <any>res2.b.date.toISOString();
+    src1.b.date = <any>src1.b.date.toISOString();
+
     src1a.a = null;
     sr.deepEqual(res1, src1);
 
-    var res1 = Objects.cloneDeep(src1b);
     src1b.c.c1 = null;
-    sr.deepEqual(res1, src1);
+    sr.deepEqual(res2, src1);
 });